work around import cycle caused by stores_test.go
